+++ Assets/XRI_Examples/SocketInteractors/Scripts/XRInfiniteInteractable.cs
@@ -1,4 +1,6 @@
 using UnityEngine.XR.Interaction.Toolkit;
+using UnityEngine.XR.Interaction.Toolkit.Interactors;
+using UnityEngine.XR.Interaction.Toolkit.Interactables;
 
 namespace UnityEngine.XR.Content.Interaction
 {