+++ Assets/XRI_Examples/UI_3D/Scripts/XRPushButton.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using UnityEngine.Events;
 using UnityEngine.XR.Interaction.Toolkit;
+using UnityEngine.XR.Interaction.Toolkit.Interactors;
+using UnityEngine.XR.Interaction.Toolkit.Interactables;
 
 namespace UnityEngine.XR.Content.Interaction
 {