+++ Assets/XRI_Examples/VRWalkthrough/Scripts/WalkthroughStep.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine.Events;
-using UnityEngine.XR.Interaction.Toolkit;
+using UnityEngine.XR.Interaction.Toolkit.Interactors;
+using UnityEngine.XR.Interaction.Toolkit.Interactables;
+using UnityEngine.XR.Interaction.Toolkit.Locomotion.Teleportation;
 
 namespace UnityEngine.XR.Content.Walkthrough
 {